+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Added global permissions for get read-access to values such as tenant’s locale, timezone, and currency. Refs UICHKOUT-983.
 * Added focus improvement for modal. Refs UICHKOUT-982.
 * Add 'subscribesTo' field for servicepoints in package.json. Refs UICHKOUT-882.
+* *BREAKING* Replace `users/configurations/entry` endpoint with `users/settings/entries` for user settings. Refs UICHKOUT-987.
 
 ## [12.0.2] (https://github.com/folio-org/ui-checkout/tree/v12.0.2) (2025-10-29)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v12.0.1...v12.0.2)