+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * *BREAKING* Replace `users/configurations/entry` endpoint with `users/settings/entries` for user settings. Refs UICHKOUT-987.
 * *BREAKING* Migrate config values to mod-configuration. Refs UICHKOUT-963.
 * *BREAKING* Add custom fields in `ViewPatron` for user and proxy user details. Refs UICHKOUT-989.
+* Add support for use circulation-bff for circulation-bff/inventory/items. Refs UICHKOUT-985.
 
 ## [12.0.2] (https://github.com/folio-org/ui-checkout/tree/v12.0.2) (2025-10-29)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v12.0.1...v12.0.2)